// frontend/src/type/type.d.ts
export interface IMenuItem {
  id: number;
  to: string;
  label: string;
}
export interface IMenuProp {
  title: string;
  data: IMenuItem[];
}
export interface IStudent {
  id: number;
  mssv: string;
  ho: string;
  ten_dem: string;
  ten: string;
  gioi_tinh: string;
  ten_lop: string;
  khoa: string;
  ten_nganh: string;
  so_dien_thoai: sting;
}